python/requirements.txt
Flask==1.1.4
azure-ai-documentintelligence==1.0.0b2
python-dotenv==0.15.0
markupsafe==2.0.1
